Move intl setup to i18n directory
