+++ app/validators/walletCreateFormValidator.js
@@ -1,7 +1,7 @@
 import MobxReactForm from 'mobx-react-form';
 import { defineMessages } from 'react-intl';
 import { createWallet } from '../actions/wallet-actions';
-import { intl } from '../i18n-setup';
+import { intl } from '../i18n';
 
 const messages = defineMessages({
   invalidWalletName: {
